python/pyproject.toml
[project]
name = "thrift-parser"
version = "0.0.3"
description = "a thrift parser (antlr4)"
authors = [
    {name = "alingse", email = "alingse@foxmail.com"},
]
dependencies = [
    "antlr4-python3-runtime>=4.10",
]
requires-python = ">=3.7"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]

[tool]
[tool.pdm]
excludes = ["tests/"]

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.1.2",
]
dev = [
    "flake8>=4.0.1",
]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
